refactor: Allow lazy subfolder plugin setup
Rather than specifying all the subfolders inside the lazy setup call,
use vim.list_extend to extend the specs in the init files.

This works easily for everything except the ai plugins. Copilot expects
Lualine's opts function to have already been called, but that isn't the
case with the basic restructure. The solution is to make a new "extras"
subfolder that is explicitly called.