fix[vscode]: Don't load plugins that are Neovim specific
Code for the color scheme and which-key should only run when in Neovim
itself. In VSCode, this causes errors, so ensure that we are _not_ in
VSCode before calling them.

It would be better if we could move the vscode specific config, rather
than doing checks in multiple files, but this works for now.