fix(python): Only close triple quotes if actually triple quotes
Previously, if an empty string was made on the line, and then an opening
quote was made for another string, it would automatically close it as if
it were a triple quote, which isn't intended.

This checks whether the cursor is actually right after the two opening
quotes before adding the close.