refactor(lang): Move lang into extras for loading order
Also adds plugins for SQL. But this comes with the complication that the
linting plugins need to be loaded before the lang ones. As plugin
loading is alphabetic within folders, by file or folder name, the
linting one was always called after lang. By moving lang to extras, the
load order is fixed.