feat(markdown): Add syntax highlighting for inline code
Allows inline code written like `{html}<p>` to be rendered using the
language inside the squiggly brackets.

Due to limitations in the way Treesitter injections work (can't match on
an arbitrary group and use that to determine the offset), and that the
markdown inline parser doesn't provide a node that is just the code
inside the code span (the `code_span` node includes the leading and
trailing delimiters), the language names need to be hardcoded. So I've
done them for html, javascript, typescript, css, python, and lua.

There are two parts to this:
1. Rendering the code with the treesitter parser, using treesitter
injections.
2. Concealing the leading part in squiggly brackets. Doing this with
treesitter didn't work, as it led to the entire node being concealed,
instead of just the language. So I used Vim syntax highlighting (just
the conceal part) with a Vim regex pattern. Due to the way that
Treesitter is loaded, it overrides this unless it is done after
Treesitter is loaded, so this is put in an ftplugin file that calls
`vim.schedule`.