fix(theme): Use flavors to correctly get variants
The lazy.nvim plugin manager loads the theme that is set for its
colorscheme automatically, regardless of light or dark mode. It needs to
be the exact colorscheme name, otherwise it defaults to the default for
that colorscheme, which is mocha for catppuccin.

So if it is loaded early (to install a new plugin) it would use the dark
theme instead of the light theme, if just told to load catppuccin. We
need to explicitly tell it to load catppuccin-latte. So the flavor needs
to be loaded.

But if we just tell it the flavor inside theme.lua, we have an issue,
because that is also used for loading the plugin, which needs the
"non-flavored" name.

So store both values inside the theme file, and load whichever is
needed.

Also uses theme.lua to populate the colors for vim lumen.