feat(yazi): Use Yazi for directories, instead of Neotree
Neotree hijacked Netrw, which is fine, but Neotree is a fallback, I'd
prefer to use Yazi for that.

Easy enough to do, and means that doing something like `nvim .` works
with Yazi.

One thing is that this means yazi.nvim can no longer be lazy loaded on
cmd or keymap, so set event to VeryLazy for it.